Fix subsystem reset issue
The problem was that some code wouldn't blink the LED properly.
After a lot of investigation, it was discovered that the core register
which was supposed to contain the address of the peripheral register
contained wrong value. Furthermore, this problem happened only
when the code was optimized and was let run freely. If the optimized
code was stepped through at the assembly level, things would work
fine. So, it looks like a race condition at a hardware level where the
peripheral maybe didn't finish coming out of reset before its registers
were accessed. Although, the true cause of this issue is still unknown
but waiting for subsystem reset to finish seems to have fixed the
problem. If you want to reproduce the issue, run tutorial 3 code with
the clocks and timers implemented, but no `.data` or `.bss` sections
initialized.